+++ netflix_clone/src/components/ContentsDetail/index.ts
@@ -1,0 +1,4 @@
+export { default as ContentsDetailModal } from './ContentsDetailModal';
+export { default as VideoPreview } from './VideoPreview';
+export { default as MovieInfo } from './MovieInfo';
+export { default as RecommendedMovies } from './RecommendedMovies';
